feat(consensus): minimmit — Phase 0 -- Crypto & committee math
Phase 0 of the HotStuff -> Minimmit migration: crypto-side quorum math
and the first consensus-side minimmit module, additive beside HotStuff.

- #511 crypto::quorum minimmit_thresholds: n >= 5f + 1 sizing with
  M = 2f + 1 (form cert => advance view) and L = n - f (finalize)
- #512 committee sizing guard rejecting sets outside minimmit bounds
- #514 dual-threshold test helper exercising M- and L-level quorums
- #546 shared crypto::QuorumCertificate signer_bitmap u64 -> u16 with
  MAX_VALIDATORS = 16; existing HotStuff QC call sites (consensus vote/
  checkpoint, chain-adapter wire/deposit/withdrawal, custody, decision-
  markets) updated in place
- #513 crates/consensus/src/minimmit/ module root + committee.rs
  (MinimmitCommittee, Certificate seam, ThresholdKind), exported
  unconditionally from consensus::lib

Closes #511 #512 #514 #546 #513

+++ crates/chain-adapter/src/deposit.rs
@@ -292,7 +292,7 @@ impl DepositCertificate {
                 need: policy.min_confirmations(),
             });
         }
-        if self.observer_bitmap != self.quorum_signature.signer_bitmap {
+        if self.observer_bitmap != u64::from(self.quorum_signature.signer_bitmap) {
             return Err(AdapterError::QuorumNotMet);
         }
         if self.quorum_signature.message != self.message_hash() {
@@ -355,7 +355,7 @@ pub fn certify_deposit(
         amount: deposit.amount,
         destination_account: deposit.destination_account,
         finality_proof: deposit.finality_proof,
-        observer_bitmap: qc.signer_bitmap,
+        observer_bitmap: u64::from(qc.signer_bitmap),
         quorum_signature: qc,
     }
 }
